refactor: move logging and tracing setup out of main.rs
setup_logs, enable_tracing and their span processor / exporter types
compose very large generic trees (tracing_subscriber layer stacks, OTLP
batch processors). Everything main.rs monomorphizes is emitted into the
binary's own codegen units, which are the last unit on the build's
critical path and overlap with nothing.

The openobserve library does not depend on openobserve-api-http, so its
unit compiles in parallel with the API crates - in the cold build it
finished 23.9s before the binary needed it. Moving the setup there puts
that work in the window instead of at the tail.

Two tests moved with the code and were dropped rather than relocated:
test_setup_logs and test_enable_tracing_error_handling both install a
global tracing subscriber, and setup_logs() calls .init() unguarded, so
they can only ever pass depending on ordering. They used to race only
with each other inside the bin test binary; keeping them would have put
that race in the lib test binary next to the cli and migration tests.
Neither asserted anything - the second wraps the call in catch_unwind
and discards the result.

main.rs 1304 -> 712 lines; the binary unit 88.8s -> 80.7s. The library
unit grows 30.9s -> 62.3s and is fully absorbed by the existing slack.
