+++ src/lib.rs
@@ -17,3 +17,4 @@
 
 pub mod cli;
 pub mod migration;
+pub mod telemetry;
