+++ src/main.rs
@@ -19,45 +19,31 @@
 
 #[cfg(feature = "tokio-console")]
 use std::net::SocketAddr;
-use std::{
-    collections::HashMap,
-    str::FromStr,
-    time::{Duration, SystemTime},
-};
+use std::time::{Duration, SystemTime};
 
 use common::{infra::cluster, meta};
 use config::{
-    META_ORG_ID,
-    cluster::LOCAL_NODE,
-    get_config,
+    META_ORG_ID, get_config,
     meta::triggers::{Trigger, TriggerModule, TriggerStatus},
     utils::size::bytes_to_human_readable,
 };
 use db::{self, scheduler::TriggerModule::QueryRecommendations};
 use infra::runtime::{create_grpc_runtime, create_job_runtime};
-use openobserve::{cli::basic::cli, migration};
+use openobserve::{
+    cli::basic::cli,
+    migration,
+    telemetry::{enable_tracing, setup_logs},
+};
 use openobserve_api_http::handler::http::router::*;
 use openobserve_core::{bootstrap, metadata};
 use openobserve_jobs::job;
-use opentelemetry::{KeyValue, global, trace::TracerProvider};
-use opentelemetry_otlp::{WithExportConfig, WithHttpConfig, WithTonicConfig};
-use opentelemetry_sdk::{Resource, propagation::TraceContextPropagator};
 use tokio::sync::oneshot;
-use tonic::{
-    codec::CompressionEncoding,
-    metadata::{MetadataKey, MetadataMap, MetadataValue},
-};
 #[cfg(feature = "enterprise")]
 use tower_http::trace::TraceLayer;
 use tracing_appender::non_blocking::WorkerGuard;
-use tracing_opentelemetry::OpenTelemetryLayer;
-use tracing_subscriber::Registry;
 use utoipa::OpenApi;
 #[cfg(feature = "enterprise")]
-use {
-    config::Config,
-    o2_enterprise::enterprise::{ai, common::config::O2Config},
-};
+use {config::Config, o2_enterprise::enterprise::common::config::O2Config};
 
 #[cfg(feature = "mimalloc")]
 #[global_allocator]
@@ -65,9 +51,6 @@ static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
 #[cfg(feature = "jemalloc")]
 #[global_allocator]
 static GLOBAL: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
-use tracing_subscriber::{
-    EnvFilter, filter::LevelFilter as TracingLevelFilter, fmt::Layer, prelude::*,
-};
 
 #[cfg(feature = "profiling")]
 #[allow(non_upper_case_globals)]
@@ -448,528 +431,6 @@ async fn main() -> Result<(), anyhow::Error> {
     Ok(())
 }
 
-/// Setup the tracing related components
-pub(crate) fn setup_logs() -> tracing_appender::non_blocking::WorkerGuard {
-    use tracing_subscriber::fmt::writer::BoxMakeWriter;
-
-    let cfg = get_config();
-    let (writer, guard) = if cfg.log.file_dir.is_empty() {
-        let (non_blocking, _guard) = tracing_appender::non_blocking(std::io::stdout());
-        (BoxMakeWriter::new(non_blocking), _guard)
-    } else {
-        let file_name_prefix = if cfg.log.file_name_prefix.is_empty() {
-            format!("o2.{}.log", cfg.common.instance_name.as_str())
-        } else {
-            cfg.log.file_name_prefix.to_string()
-        };
-        let file_appender = tracing_appender::rolling::daily(&cfg.log.file_dir, file_name_prefix);
-        let (non_blocking, _guard) = tracing_appender::non_blocking(file_appender);
-        (BoxMakeWriter::new(non_blocking), _guard)
-    };
-    let layer = if cfg.log.json_format {
-        Layer::default()
-            .with_writer(writer)
-            .with_timer(config::meta::logger::CustomTimeFormat)
-            .with_ansi(false)
-            .json()
-            .with_current_span(false)
-            .with_span_list(false)
-            .boxed()
-    } else {
-        Layer::default()
-            .with_writer(writer)
-            .with_ansi(false)
-            .with_target(true)
-            .event_format(config::meta::logger::O2Formatter::default())
-            .boxed()
-    };
-
-    tracing_subscriber::registry()
-        .with(
-            EnvFilter::builder()
-                .with_default_directive(TracingLevelFilter::INFO.into())
-                .from_env_lossy(),
-        )
-        .with(layer)
-        .init();
-    guard
-}
-
-/// Custom span processor that filters spans based on their name prefix
-#[cfg(feature = "enterprise")]
-#[derive(Debug)]
-struct FilteringSpanProcessor<P> {
-    inner: P,
-    prefix_filter: Option<String>,
-}
-
-#[cfg(feature = "enterprise")]
-impl<P> FilteringSpanProcessor<P> {
-    fn new(inner: P, prefix_filter: Option<String>) -> Self {
-        Self {
-            inner,
-            prefix_filter,
-        }
-    }
-}
-
-#[cfg(feature = "enterprise")]
-impl<P: opentelemetry_sdk::trace::SpanProcessor> opentelemetry_sdk::trace::SpanProcessor
-    for FilteringSpanProcessor<P>
-{
-    fn on_start(&self, span: &mut opentelemetry_sdk::trace::Span, cx: &opentelemetry::Context) {
-        // Always call inner on_start - we can only filter on_end when we have the full span data
-        self.inner.on_start(span, cx);
-    }
-
-    fn on_end(&self, span: opentelemetry_sdk::trace::SpanData) {
-        // If no filter is set, pass through all spans
-        if let Some(prefix) = &self.prefix_filter {
-            let span_name = span.name.as_ref();
-            // Only process spans that match the prefix
-            if !span_name.starts_with(prefix) {
-                return;
-            }
-        }
-        self.inner.on_end(span);
-    }
-
-    fn force_flush(&self) -> Result<(), opentelemetry_sdk::error::OTelSdkError> {
-        self.inner.force_flush()
-    }
-
-    fn shutdown(&self) -> Result<(), opentelemetry_sdk::error::OTelSdkError> {
-        self.inner.shutdown()
-    }
-
-    fn shutdown_with_timeout(
-        &self,
-        timeout: std::time::Duration,
-    ) -> Result<(), opentelemetry_sdk::error::OTelSdkError> {
-        self.inner.shutdown_with_timeout(timeout)
-    }
-}
-
-/// Custom SpanExporter that ingests traces directly into the _meta org.
-/// This is used by the search inspector feature to ensure search profiling
-/// traces are always available in the _meta org, regardless of where the user's
-/// OTLP endpoint points.
-///
-/// Handles both single-node and multi-node deployments:
-/// - Ingester nodes: calls `handle_otlp_request` directly
-/// - Non-ingester nodes (e.g. queriers): forwards via gRPC to an ingester
-#[derive(Debug)]
-struct MetaOrgTraceExporter {
-    resource: opentelemetry_proto::transform::common::tonic::ResourceAttributesWithSchema,
-}
-
-impl MetaOrgTraceExporter {
-    fn new(
-        resource: opentelemetry_proto::transform::common::tonic::ResourceAttributesWithSchema,
-    ) -> Self {
-        Self { resource }
-    }
-}
-
-impl opentelemetry_sdk::trace::SpanExporter for MetaOrgTraceExporter {
-    fn export(
-        &self,
-        batch: Vec<opentelemetry_sdk::trace::SpanData>,
-    ) -> impl std::future::Future<Output = opentelemetry_sdk::error::OTelSdkResult> + Send {
-        // Convert SpanData to ExportTraceServiceRequest synchronously
-        // so we don't hold a borrow on &self across await points
-        let resource_spans =
-            opentelemetry_proto::transform::trace::tonic::group_spans_by_resource_and_scope(
-                batch,
-                &self.resource,
-            );
-        let request = opentelemetry_proto::tonic::collector::trace::v1::ExportTraceServiceRequest {
-            resource_spans,
-        };
-
-        async move {
-            if LOCAL_NODE.is_ingester() {
-                // Ingest directly on ingester nodes
-                match openobserve_core::traces::handle_otlp_request(
-                    META_ORG_ID,
-                    request,
-                    config::meta::otlp::OtlpRequestType::HttpJson,
-                    None,
-                    ingestion_common::IngestUser::SystemJob(
-                        ingestion_common::SystemJobType::SelfReporting,
-                    ),
-                )
-                .await
-                {
-                    Ok(resp) if resp.status().is_success() => {
-                        log::debug!("[SEARCH-INSPECTOR] Successfully ingested traces to _meta org");
-                        Ok(())
-                    }
-                    Ok(resp) => {
-                        log::error!(
-                            "[SEARCH-INSPECTOR] Error ingesting traces to _meta org: status={}",
-                            resp.status()
-                        );
-                        Err(opentelemetry_sdk::error::OTelSdkError::InternalFailure(
-                            format!("Failed to ingest traces: status={}", resp.status()),
-                        ))
-                    }
-                    Err(e) => {
-                        log::error!("[SEARCH-INSPECTOR] Error ingesting traces to _meta org: {e}");
-                        Err(opentelemetry_sdk::error::OTelSdkError::InternalFailure(
-                            format!("Failed to ingest traces: {e}"),
-                        ))
-                    }
-                }
-            } else {
-                // Forward to an ingester node via gRPC
-                let cfg = get_config();
-                let token = match config::meta::cluster::get_internal_grpc_token()
-                    .parse::<tonic::metadata::MetadataValue<_>>()
-                {
-                    Ok(token) => token,
-                    Err(e) => {
-                        log::error!("[SEARCH-INSPECTOR] Failed to parse internal gRPC token: {e}");
-                        return Err(opentelemetry_sdk::error::OTelSdkError::InternalFailure(
-                            format!("Failed to parse gRPC token: {e}"),
-                        ));
-                    }
-                };
-
-                let (_addr, channel) = match openobserve_node::grpc::get_ingester_channel().await {
-                    Ok(v) => v,
-                    Err(e) => {
-                        log::error!("[SEARCH-INSPECTOR] Failed to get ingester channel: {e}");
-                        return Err(opentelemetry_sdk::error::OTelSdkError::InternalFailure(
-                            format!("No ingester available: {e}"),
-                        ));
-                    }
-                };
-
-                let client = opentelemetry_proto::tonic::collector::trace::v1::trace_service_client::TraceServiceClient::with_interceptor(
-                    channel,
-                    move |mut req: tonic::Request<()>| {
-                        req.metadata_mut().insert("authorization", token.clone());
-                        Ok(req)
-                    },
-                );
-
-                let org_header_key: MetadataKey<_> = cfg.grpc.org_header_key.parse().unwrap();
-                let mut grpc_request = tonic::Request::new(request);
-                grpc_request.metadata_mut().insert(
-                    org_header_key,
-                    META_ORG_ID
-                        .parse::<tonic::metadata::MetadataValue<_>>()
-                        .unwrap(),
-                );
-
-                match client
-                    .send_compressed(CompressionEncoding::Gzip)
-                    .accept_compressed(CompressionEncoding::Gzip)
-                    .max_decoding_message_size(cfg.grpc.max_message_size * 1024 * 1024)
-                    .max_encoding_message_size(cfg.grpc.max_message_size * 1024 * 1024)
-                    .export(grpc_request)
-                    .await
-                {
-                    Ok(_) => {
-                        log::debug!(
-                            "[SEARCH-INSPECTOR] Successfully forwarded traces to ingester for _meta org"
-                        );
-                        Ok(())
-                    }
-                    Err(e) => {
-                        log::error!("[SEARCH-INSPECTOR] Error forwarding traces to ingester: {e}");
-                        Err(opentelemetry_sdk::error::OTelSdkError::InternalFailure(
-                            format!("Failed to forward traces to ingester: {e}"),
-                        ))
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn enable_tracing() -> Result<opentelemetry_sdk::trace::SdkTracerProvider, anyhow::Error> {
-    let cfg = get_config();
-    opentelemetry::global::set_text_map_propagator(TraceContextPropagator::new());
-
-    let mut tracer_builder = opentelemetry_sdk::trace::SdkTracerProvider::builder();
-
-    // Add main OpenObserve OTLP exporter (if general tracing is enabled)
-    if cfg.common.tracing_enabled || cfg.common.tracing_search_enabled {
-        tracer_builder = if cfg.common.otel_otlp_grpc_url.is_empty() {
-            tracer_builder.with_span_processor(
-            opentelemetry_sdk::trace::span_processor_with_async_runtime::BatchSpanProcessor::builder(
-                {
-                    let mut headers = HashMap::new();
-                    headers.insert(
-                        cfg.common.tracing_header_key.clone(),
-                        cfg.common.tracing_header_value.clone(),
-                    );
-                    opentelemetry_otlp::SpanExporter::builder()
-                        .with_http()
-                        .with_http_client(
-                            reqwest::Client::builder()
-                        .danger_accept_invalid_certs(true)
-                        .timeout(Duration::from_secs(10))           // Overall request timeout
-                        .connect_timeout(Duration::from_secs(5))    // Connection establishment timeout
-                        .pool_idle_timeout(Duration::from_secs(60)) // How long to keep idle connections
-                        .pool_max_idle_per_host(10) // How many idle connections to keep per host
-                        .build()?,
-                        )
-                        .with_endpoint(&cfg.common.otel_otlp_url)
-                        .with_headers(headers)
-                        .build()?
-                },
-                opentelemetry_sdk::runtime::Tokio,
-            )
-            .build(),
-        )
-        } else {
-            tracer_builder.with_span_processor(
-            opentelemetry_sdk::trace::span_processor_with_async_runtime::BatchSpanProcessor::builder(
-                {
-                    let mut metadata = MetadataMap::new();
-                    metadata.insert(
-                        MetadataKey::from_str(&cfg.common.tracing_header_key).unwrap(),
-                        MetadataValue::from_str(&cfg.common.tracing_header_value).unwrap(),
-                    );
-                    metadata.insert(
-                        MetadataKey::from_str(&cfg.grpc.org_header_key).unwrap(),
-                        MetadataValue::from_str(&cfg.common.tracing_grpc_header_org).unwrap(),
-                    );
-                    metadata.insert(
-                        MetadataKey::from_str(&cfg.grpc.stream_header_key).unwrap(),
-                        MetadataValue::from_str(&cfg.common.tracing_grpc_header_stream_name)
-                            .unwrap(),
-                    );
-                    opentelemetry_otlp::SpanExporter::builder()
-                        .with_tonic()
-                        .with_endpoint(&cfg.common.otel_otlp_grpc_url)
-                        .with_metadata(metadata)
-                        .with_protocol(opentelemetry_otlp::Protocol::Grpc)
-                        .build()?
-                },
-                opentelemetry_sdk::runtime::Tokio,
-            )
-            .build(),
-        )
-        };
-        log::info!("Main OpenObserve OTLP exporter configured");
-    }
-
-    // Handle AI tracing (enterprise feature)
-    #[cfg(feature = "enterprise")]
-    {
-        use o2_enterprise::enterprise::common::config::get_config as get_o2_config;
-        let o2_cfg = get_o2_config();
-
-        // If AI tracing is enabled but general tracing is NOT enabled,
-        // we need to add OpenObserve OTLP exporter for AI traces
-        if o2_cfg.ai.tracing_enabled
-            && !cfg.common.tracing_enabled
-            && !cfg.common.tracing_search_enabled
-        {
-            log::info!("AI tracing enabled independently - configuring OpenObserve OTLP exporter");
-
-            // Add OpenObserve exporter for AI traces
-            if !cfg.common.otel_otlp_url.is_empty() {
-                let mut headers = HashMap::new();
-                headers.insert(
-                    cfg.common.tracing_header_key.clone(),
-                    cfg.common.tracing_header_value.clone(),
-                );
-
-                let oo_exporter = opentelemetry_otlp::SpanExporter::builder()
-                    .with_http()
-                    .with_http_client(
-                        reqwest::Client::builder()
-                            .danger_accept_invalid_certs(true)
-                            .timeout(Duration::from_secs(10))
-                            .connect_timeout(Duration::from_secs(5))
-                            .pool_idle_timeout(Duration::from_secs(60))
-                            .pool_max_idle_per_host(10)
-                            .build()?,
-                    )
-                    .with_endpoint(&cfg.common.otel_otlp_url)
-                    .with_headers(headers)
-                    .build()?;
-
-                let oo_processor = opentelemetry_sdk::trace::span_processor_with_async_runtime::BatchSpanProcessor::builder(
-                    oo_exporter,
-                    opentelemetry_sdk::runtime::Tokio,
-                )
-                .build();
-
-                // Wrap with filtering processor to only send AI traces
-                let filtered_processor = FilteringSpanProcessor::new(
-                    oo_processor,
-                    Some("ai.".to_string()), // Only send spans starting with "ai."
-                );
-
-                tracer_builder = tracer_builder.with_span_processor(filtered_processor);
-                log::info!(
-                    "AI traces (ai.* spans only) will be sent to OpenObserve: {}",
-                    cfg.common.otel_otlp_url
-                );
-            } else if !cfg.common.otel_otlp_grpc_url.is_empty() {
-                let mut metadata = MetadataMap::new();
-                metadata.insert(
-                    MetadataKey::from_str(&cfg.common.tracing_header_key).unwrap(),
-                    MetadataValue::from_str(&cfg.common.tracing_header_value).unwrap(),
-                );
-
-                let oo_exporter = opentelemetry_otlp::SpanExporter::builder()
-                    .with_tonic()
-                    .with_endpoint(&cfg.common.otel_otlp_grpc_url)
-                    .with_metadata(metadata)
-                    .with_protocol(opentelemetry_otlp::Protocol::Grpc)
-                    .build()?;
-
-                let oo_processor = opentelemetry_sdk::trace::span_processor_with_async_runtime::BatchSpanProcessor::builder(
-                    oo_exporter,
-                    opentelemetry_sdk::runtime::Tokio,
-                )
-                .build();
-
-                // Wrap with filtering processor to only send AI traces
-                let filtered_processor = FilteringSpanProcessor::new(
-                    oo_processor,
-                    Some("ai.".to_string()), // Only send spans starting with "ai."
-                );
-
-                tracer_builder = tracer_builder.with_span_processor(filtered_processor);
-                log::info!(
-                    "AI traces (ai.* spans only) will be sent to OpenObserve (gRPC): {}",
-                    cfg.common.otel_otlp_grpc_url
-                );
-            } else {
-                log::warn!(
-                    "AI tracing enabled but ZO_OTEL_OTLP_URL not configured - AI traces will not be exported"
-                );
-            }
-        }
-
-        // Additionally, if O2_AI_EVAL_OTLP_ENDPOINT is set, send AI traces to evaluation platform
-        // too
-        let eval_endpoint = std::env::var("O2_AI_EVAL_OTLP_ENDPOINT")
-            .ok()
-            .or_else(|| o2_cfg.ai.eval_otlp_endpoint.clone())
-            .filter(|s| !s.is_empty());
-
-        if let Some(endpoint) = eval_endpoint {
-            log::info!(
-                "Configuring additional AI evaluation OTLP exporter to: {}",
-                endpoint
-            );
-
-            let eval_exporter = opentelemetry_otlp::SpanExporter::builder()
-                .with_tonic()
-                .with_endpoint(endpoint)
-                .with_timeout(std::time::Duration::from_secs(10))
-                .build()?;
-
-            let eval_processor = opentelemetry_sdk::trace::span_processor_with_async_runtime::BatchSpanProcessor::builder(
-                eval_exporter,
-                opentelemetry_sdk::runtime::Tokio,
-            )
-            .build();
-
-            // Wrap with filtering processor to only send AI traces
-            let filtered_eval_processor =
-                FilteringSpanProcessor::new(eval_processor, Some("ai.".to_string()));
-
-            tracer_builder = tracer_builder.with_span_processor(filtered_eval_processor);
-            log::info!(
-                "AI evaluation OTLP exporter configured - AI traces (ai.* spans only) will be sent to evaluation platform"
-            );
-        }
-    }
-
-    // Build resource attributes (base + extra envs)
-    let mut resource_attrs = vec![
-        KeyValue::new("service.name", cfg.common.node_role.to_string()),
-        KeyValue::new("service.instance", cfg.common.instance_name.to_string()),
-        KeyValue::new("service.version", config::VERSION),
-    ];
-    for env_name in cfg
-        .common
-        .tracing_extra_envs
-        .split(',')
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-    {
-        if let Ok(val) = std::env::var(env_name)
-            && !val.is_empty()
-        {
-            resource_attrs.push(KeyValue::new(env_name.to_lowercase(), val));
-        }
-    }
-
-    // If search inspector is enabled, add a span processor that ingests traces
-    // directly into the _meta org. This ensures search profiling data is always
-    // available in _meta regardless of where the user's OTLP endpoint points.
-    if cfg.common.search_inspector_enabled {
-        let resource = Resource::builder()
-            .with_attributes(resource_attrs.clone())
-            .build();
-        let resource_attrs =
-            opentelemetry_proto::transform::common::tonic::ResourceAttributesWithSchema::from(
-                &resource,
-            );
-        let meta_exporter = MetaOrgTraceExporter::new(resource_attrs);
-        let meta_processor =
-            opentelemetry_sdk::trace::span_processor_with_async_runtime::BatchSpanProcessor::builder(
-                meta_exporter,
-                opentelemetry_sdk::runtime::Tokio,
-            )
-            .build();
-        tracer_builder = tracer_builder.with_span_processor(meta_processor);
-        log::info!(
-            "Search inspector OTLP exporter configured - traces will be ingested to _meta org"
-        );
-    }
-
-    // Add UUID v7 ID generator and resource attributes
-    tracer_builder = tracer_builder.with_id_generator({
-        #[cfg(feature = "enterprise")]
-        {
-            ai::agent::tracing::UuidV7IdGenerator
-        }
-        #[cfg(not(feature = "enterprise"))]
-        {
-            opentelemetry_sdk::trace::RandomIdGenerator::default()
-        }
-    });
-
-    let tracer =
-        tracer_builder.with_resource(Resource::builder().with_attributes(resource_attrs).build());
-
-    // build
-    let tracer = tracer.build();
-
-    let layer = if cfg.log.json_format {
-        tracing_subscriber::fmt::layer()
-            .with_ansi(false)
-            .json()
-            .boxed()
-    } else {
-        tracing_subscriber::fmt::layer().with_ansi(false).boxed()
-    };
-
-    global::set_tracer_provider(tracer.clone());
-    Registry::default()
-        .with(tracing_subscriber::EnvFilter::new(&cfg.log.level))
-        .with(layer)
-        .with(OpenTelemetryLayer::new(
-            tracer.tracer("tracing-otel-subscriber"),
-        ))
-        .init();
-
-    // Return the tracer provider
-    Ok(tracer)
-}
-
 #[cfg(feature = "enterprise")]
 async fn init_action_server() -> Result<(), anyhow::Error> {
     let cfg = get_config();
@@ -1112,46 +573,7 @@ fn check_ratelimit_config(cfg: &Config, o2cfg: &O2Config) -> Result<(), anyhow::
 
 #[cfg(test)]
 mod tests {
-    use tokio::runtime::Runtime;
-
     use super::*;
-
-    #[test]
-    fn test_setup_logs() {
-        let _guard = setup_logs();
-
-        // Just verify that the guard is valid and the logs setup doesn't panic
-    }
-
-    #[test]
-    fn test_enable_tracing_error_handling() {
-        // Test that enable_tracing handles configuration errors gracefully
-        // This test verifies the function exists and can be called
-        // In a real environment, tracing setup might fail due to network issues
-
-        // We can't easily test the actual tracing setup without mocking external services
-        // But we can ensure the function signature and basic error handling work
-        let result = std::panic::catch_unwind(|| {
-            // Just verify the function can be called
-            let rt = Runtime::new().unwrap();
-            rt.block_on(async {
-                // This might fail in test environment due to missing config
-                // but that's expected and we're testing error handling
-                let _ = enable_tracing();
-            });
-        });
-
-        // The test should pass regardless of whether enable_tracing() succeeds or fails
-        // In test environments, it may fail due to:
-        // 1. Global subscriber already set by another test (when running in parallel)
-        // 2. Missing configuration
-        // 3. Network issues
-        // We're testing that it doesn't panic unexpectedly beyond expected tracing setup issues
-        // Don't assert result.is_ok() because parallel tests will fail due to global subscriber
-        // conflicts (expected in test environment when tests run in parallel)
-        // The important thing is that we can call the function without unexpected panics
-        let _ = result;
-    }
 
     #[cfg(feature = "enterprise")]
     #[test]
@@ -1257,20 +679,6 @@ mod tests {
         assert!(!wait_for_send);
         assert!(server_stop);
         assert!(wait_for_stop);
-    }
-
-    #[test]
-    fn test_resource_key_value_creation() {
-        use opentelemetry::KeyValue;
-
-        // Test KeyValue creation patterns used in tracing setup
-        let service_name = KeyValue::new("service.name", "test-service");
-        let service_instance = KeyValue::new("service.instance", "test-instance");
-        let service_version = KeyValue::new("service.version", "1.0.0");
-
-        assert_eq!(service_name.key.as_str(), "service.name");
-        assert_eq!(service_instance.key.as_str(), "service.instance");
-        assert_eq!(service_version.key.as_str(), "service.version");
     }
 
     #[tokio::test]
